requirements.txt
fastapi
uvicorn[standard]
chronos-forecasting
psutil
